refactor(vscode): dedupe DOM-lookup helpers into webview/shared/domRefs
`requireElementById` was duplicated between `preview/behavior.ts` and
`history/behavior.ts` — same fail-fast `document.getElementById` body,
same throw-on-miss semantics. Moves both helpers (+ `requireSelector`
which only the preview side currently uses) to `webview/shared/domRefs.ts`
so the two panels share one source of truth and any future panel can
pick them up without re-implementing.

`preview/behavior.ts`: 542 → 527 lines (−15 LOC).
`history/behavior.ts`: 253 → 245 lines (−8 LOC).
`webview/shared/domRefs.ts`: +34 lines (new file).

Net +11 LOC because the new file's jsdoc balances the extraction, but
the helper now has a single source of truth — the dedup goal — and the
behaviour stays identical.

Drive-by audit while picking the next slice: `grep -rEn "@ts-nocheck|
@ts-ignore|@ts-expect-error|as any|: any" src/` returns zero
non-comment matches across the host + webview tree. The `@ts-nocheck`
removal arc that started after #767 is fully closed.

## Test plan

- [x] `npm run compile` clean (host + webview)
- [x] `npm test` — 421/421 passing
- [x] `npm run format:check` clean
- [ ] Manual smoke: open both webviews, verify the static template ids
  are still found (`#preview-grid`, `#focus-controls`, `#timeline`,
  `#filter-source`, etc.).
